refactor(core): Mark worker-pools files with the .ee suffix (no-changelog)
Worker pools are gated behind an enterprise license, so the
implementation files move under the Enterprise License per LICENSE.md.
Pure renames plus import-path updates; no logic changes.
