+++ packages/@n8n/db/src/entities/project-pool-settings.ee.ts
