+++ packages/cli/src/scaling/worker-pools.service.ee.ts
