+++ packages/cli/src/workflow-runner.ts
@@ -50,7 +50,7 @@ import { ExternalHooks } from '@/external-hooks';
 import type { ResumableExecution } from '@/interfaces';
 import { ManualExecutionService } from '@/manual-execution.service';
 import { NodeTypes } from '@/node-types';
-import type { PoolConfigService } from '@/scaling/pool-config.service';
+import type { PoolConfigService } from '@/scaling/pool-config.service.ee';
 import type { ScalingService } from '@/scaling/scaling.service';
 import type { Job, JobData } from '@/scaling/scaling.types';
 import { EngineV2Dispatcher } from '@/services/engine-v2-dispatcher.service';
@@ -550,7 +550,7 @@ export class WorkflowRunner {
 		}
 
 		if (!this.poolConfigService) {
-			const { PoolConfigService } = await import('@/scaling/pool-config.service.js');
+			const { PoolConfigService } = await import('@/scaling/pool-config.service.ee.js');
 			this.poolConfigService = Container.get(PoolConfigService);
 		}
